Wrap grid for refactoring
